Move read to its own file.
Move the read proc to its own file read.js, and move the common
module-specific stuff to index.js. The write proc will go into
write.js.
