+++ lib/read.js
@@ -3,7 +3,6 @@ var JuttleErrors = require('juttle/lib/errors');
 var JuttleMoment = require('juttle/lib/moment/juttle-moment');
 var Promise = require('bluebird');
 var google = require('googleapis');
-var googleAuth = require('google-auth-library');
 var _ = require('underscore');
 var Batchelor = require('batchelor');
 var FilterGmailCompiler = require('./filter-gmail-compiler');
@@ -255,37 +254,12 @@ var Read = Juttle.proc.source.extend({
 
 });
 
-// Non-juttle pluggable adapter api functions below here.
-// Taken from the quickstart page:
-// https://developers.google.com/gmail/api/quickstart/nodejs, and
-// then converted to use promises.
-
-/**
- * Return an OAuth2 client with the given credentials and token.
- *
- * @param {Object} credentials The authorization client credentials.
- * @param {Object} token An OAuth2 token.
- */
-function authorize(credentials, token) {
-    var clientSecret = credentials.installed.client_secret;
-    var clientId = credentials.installed.client_id;
-    var redirectUrl = credentials.installed.redirect_uris[0];
-    var auth = new googleAuth();
-    var oauth2Client = new auth.OAuth2(clientId, clientSecret, redirectUrl);
-
-    oauth2Client.credentials = token;
-    return oauth2Client;
+function init(provided_auth) {
+    auth = provided_auth;
 }
 
-var GmailBackend = function(config) {
-
-    auth = authorize(config["client-credentials"],
-                     config["oauth2-token"]);
-
-    return {
-        name: 'gmail',
-        read: Read
-    };
+module.exports = {
+    init: init,
+    read: Read
 };
 
-module.exports = GmailBackend;
